Update RequirementsList_Library.xml
Open this in excel or sheets or numbers or something if you aren't busy. Even if you can't mod, I still want your input on what you want added the most. Under relative rating, put 5 if you want it in, but not that much, and put 10 if it must be in the mod or it isn't worth doing, like animations. It's an xml so it could be uploaded to github.